Add appropriate 64-bit mac flags for src/media targets and fix a compilation error.
BUG=150980


Review URL: https://chromiumcodereview.appspot.com/10970007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@158185 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/audio/mac/audio_manager_mac.cc
@@ -32,7 +32,7 @@ static bool HasAudioHardware(AudioObjectPropertySelector selector) {
     kAudioObjectPropertyScopeGlobal,            // mScope
     kAudioObjectPropertyElementMaster           // mElement
   };
-  size_t output_device_id_size = sizeof(output_device_id);
+  UInt32 output_device_id_size = static_cast<UInt32>(sizeof(output_device_id));
   OSStatus err = AudioObjectGetPropertyData(kAudioObjectSystemObject,
                                             &property_address,
                                             0,     // inQualifierDataSize
